fix(category):Add tag delete check

+++ weblog-admin/src/main/java/com/jacob/weblog/admin/service/impl/AdminTagServiceImpl.java
@@ -1,12 +1,16 @@
 package com.jacob.weblog.admin.service.impl;
 
+import com.baomidou.mybatisplus.core.toolkit.Wrappers;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.jacob.weblog.admin.model.vo.tag.*;
 import com.jacob.weblog.admin.service.AdminTagService;
+import com.jacob.weblog.common.domain.dos.ArticleTagRelDO;
 import com.jacob.weblog.common.domain.dos.TagDO;
+import com.jacob.weblog.common.domain.mapper.ArticleTagRelMapper;
 import com.jacob.weblog.common.domain.mapper.TagMapper;
 import com.jacob.weblog.common.enums.ResponseCodeEnum;
+import com.jacob.weblog.common.exception.BizException;
 import com.jacob.weblog.common.model.vo.SelectRspVO;
 import com.jacob.weblog.common.utils.PageResponse;
 import com.jacob.weblog.common.utils.Response;
@@ -18,6 +22,7 @@ import org.springframework.util.CollectionUtils;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 /**
@@ -31,6 +36,9 @@ import java.util.stream.Collectors;
 public class AdminTagServiceImpl extends ServiceImpl<TagMapper, TagDO> implements AdminTagService {
     @Autowired
     private TagMapper tagMapper;
+
+    @Autowired
+    private ArticleTagRelMapper articleTagRelMapper;
 
     /**
      * 添加标签集合
@@ -103,6 +111,14 @@ public class AdminTagServiceImpl extends ServiceImpl<TagMapper, TagDO> implement
         // 标签 ID
         Long tagId = deleteTagReqVO.getId();
 
+        // 校验该标签下是否有关联的文章，若有，则不允许删除，提示用户需要先删除标签下的文章
+        ArticleTagRelDO articleTagRelDO = articleTagRelMapper.selectOneByTagId(tagId);
+
+        if (Objects.nonNull(articleTagRelDO)) {
+            log.warn("==> 此标签下包含文章，无法删除，tagId: {}", tagId);
+            throw new BizException(ResponseCodeEnum.TAG_CAN_NOT_DELETE);
+        }
+
         // 根据标签 ID 删除
         int count = tagMapper.deleteById(tagId);
 
@@ -123,6 +139,25 @@ public class AdminTagServiceImpl extends ServiceImpl<TagMapper, TagDO> implement
         List<TagDO> tagDOS = tagMapper.selectByKey(key);
 
         // do 转 vo
+        List<SelectRspVO> vos = null;
+        if (!CollectionUtils.isEmpty(tagDOS)) {
+            vos = tagDOS.stream()
+                    .map(tagDO -> SelectRspVO.builder()
+                            .label(tagDO.getName())
+                            .value(tagDO.getId())
+                            .build())
+                    .collect(Collectors.toList());
+        }
+
+        return Response.success(vos);
+    }
+
+    @Override
+    public Response findTagSelectList() {
+        // 查询所有标签, Wrappers.emptyWrapper() 表示查询条件为空
+        List<TagDO> tagDOS = tagMapper.selectList(Wrappers.emptyWrapper());
+
+        // DO 转 VO
         List<SelectRspVO> vos = null;
         if (!CollectionUtils.isEmpty(tagDOS)) {
             vos = tagDOS.stream()
